actualizacion de atributos en objetos
