--- RetoG2/src/main/java/modelo/Library.java
@@ -1,10 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package modelo;
-
-public class Library {
-    public boolean someLibraryMethod() {
-        return true;
-    }
-}
